fix line number display
the line numbers are limited to unsigned short in the parser and
consequently the evaluator. as that's an internal detail that could
change any time, the display works with ints. so the special value ~0
needs to be converted to -1 in the right place.

+++ src/shared/proparser/qmakeevaluator.h
@@ -248,9 +248,9 @@ public:
 
     struct Location {
         Location() : pro(0), line(0) {}
-        Location(ProFile *_pro, int _line) : pro(_pro), line(_line) {}
+        Location(ProFile *_pro, ushort _line) : pro(_pro), line(_line) {}
         ProFile *pro;
-        int line;
+        ushort line;
     };
 
     Location m_current; // Currently evaluated location
